CI: add and run codespell
Typos happen. Some are easy to spot in a review, but other ones are good
in hiding from the human eye.
Machines can not be tricked in the same way, which is why codespell finds
some typos a human reviewer wouldn't.

Run codespell once, fix all the found typos and add it to the CI.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>